+++ tests/kernel/test_kernel_core.py
@@ -211,42 +211,7 @@ class Test_main_computation():
         print("zmax computed   : ", numpy.array2string(zmaxc, floatmode = 'unique'))
         print("zmin reference  : ", numpy.array2string(zmin, floatmode = 'unique'))
         print("zmin computed   : ", numpy.array2string(zminc, floatmode = 'unique'))
-        
-class Test_celerity_calculation():
-    
-    def test_given_first_element_Of_celq_is_smaller_than_q_when_get_celerity_then_return_first_element_of_celc(self):
-        
-        first_element_of_celc = 10  
-        q = 11.0
-        first_element_of_celq = q-1
-        cel_q = [first_element_of_celq,20,30,40] 
-        cel_c = [first_element_of_celc,20,30,40] 
-        
-        celerity = dfastmi.kernel.core.get_celerity(q, cel_q, cel_c)
-        assert celerity == 11
-    
-    def test_given_first_element_Of_celq_is_bigger_than_q_when_get_celerity_then_return_first_element_of_celc(self):
-        
-        first_element_of_celc = 10  
-        q = 11.0
-        first_element_of_celq = q+1
-        cel_q = [first_element_of_celq,20,30,40] 
-        cel_c = [first_element_of_celc,20,30,40] 
-        
-        celerity = dfastmi.kernel.core.get_celerity(q, cel_q, cel_c)
-        assert celerity == first_element_of_celc
 
-    def test_given_q_bigger_than_any_celq_when_get_celerity_then_return_last_element_of_celc(self):
-        
-        LastElementOfCelc = 40
-        
-        q = 50.0
-        cel_q = [10,20,30,40]
-        cel_c = [10,20,30,LastElementOfCelc] 
-        
-        celerity = dfastmi.kernel.core.get_celerity(q, cel_q, cel_c)
-        assert celerity == LastElementOfCelc
-    
 class Test_relax_factors_calculation():
     def test_given_single_value_for_calculation_when_relax_factors_then_return_rsigma_value_between_expected_values(self):
         Q = [2]
